# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.AuthResponse:
    properties:
      access_token:
        description: AccessToken is the JWT access token
        type: string
      expires_in:
        description: ExpiresIn is the token expiration time in seconds
        type: integer
      refresh_token:
        description: RefreshToken is the JWT refresh token
        type: string
      token_type:
        description: TokenType is the type of token (Bearer)
        type: string
      user:
        allOf:
        - $ref: '#/definitions/dto.User'
        description: User is the authenticated user information
    type: object
...
  dto.ChangePasswordRequest:
    properties:
      current_password:
        description: CurrentPassword is the user's current password for verification
        type: string
      new_password:
        description: NewPassword is the new password the user wants to set
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        description: Password is the user's password
        type: string
      phone_number:
        description: PhoneNumber is the Ethiopian phone number (+251...)
        type: string
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        description: RefreshToken is the JWT refresh token
        type: string
    type: object
  dto.RegisterUser:
    properties:
      lastname:
        description: LastName is the last name of the user.
        type: string
      name:
        description: Name is the first name of the user.
        type: string
      password:
        description: Password is the user's password.
        type: string
      phone_number:
        description: PhoneNumber is the Ethiopian phone number (+251...).
        type: string
      role:
        description: Role is the user's role (admin, manager, user) - defaults to
          'user'
        type: string
      telegram_id:
        description: TelegramID is the optional Telegram ID of the user.
        type: string
    type: object
...
  dto.User:
    properties:
      childrens_name:
        description: ChildrensName is a list of children's names
        items:
          type: string
        type: array
      created_at:
        description: |-
          CreatedAt is the time when the user is created.
          It is automatically set when the user is created.
        type: string
      deleted_at:
        description: DeletedAt is the time the user was deleted.
        type: string
      education:
        description: Education is the user's education background
        type: string
      id:
        description: |-
          ID is the unique identifier of the user.
          It is automatically generated when the user is created.
        type: string
      job_title:
        description: JobTitle is the user's job title
        type: string
      lastname:
        description: LastName is the last name of the user.
        type: string
      marriage_status:
        description: MarriageStatus is the user's marriage status
        type: string
      name:
        description: Name is the first name of the user.
        type: string
      partner_name:
        description: PartnerName is the name of the user's partner (spouse)
        type: string
      phone_number:
        description: PhoneNumber is the Ethiopian phone number of the user.
        type: string
      role:
        description: Role is the user's role (admin, manager, user)
        type: string
      status:
        description: Status is the user's account status (PENDING, ACTIVE, INACTIVE)
        type: string
      telegram_id:
        description: TelegramID is the optional Telegram ID of the user.
        type: string
      updated_at:
        description: UpdatedAt is the time the user was last updated.
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  model.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive authentication API supporting Ethiopian phone numbers
    with JWT-based security
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kesis Ethiopian Phone Authentication API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the authenticated user's password. Requires the current
        password for verification.
      parameters:
      - description: Current and new password details
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully changed password
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Bad request - validation errors or incorrect current password
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Authentication
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 'Authenticate with Ethiopian phone number and password. Returns
        JWT access and refresh tokens. Supports multiple phone formats: +251912345678,
        0912345678, or 912345678.'
      parameters:
      - description: Login credentials (phone number + password)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated - returns access token (15min) and
            refresh token (7 days)
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad request - validation errors
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid phone number or password
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Login with Ethiopian phone number
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token. The refresh
        token remains valid for its full 7-day lifetime.
      parameters:
      - description: Refresh token request
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed - returns new access token (15min) with
            same refresh token
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad request - missing or malformed refresh token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - invalid or expired refresh token
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: 'Register a new user with Ethiopian phone number. Supports multiple
        formats: +251912345678, 0912345678, or 912345678. Phone numbers are automatically
        normalized to E.164 format. Users default to ''user'' role.'
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Bad request - validation errors or duplicate phone number
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"